Add pkg-config to build image
Right now, the base build image does not include `pkg-config` which is a very common tool used by build systems to check for the existence of C libraries on the system & find out critical information like where libraries and headers live. I propose we should add this to the builder image so that tools can use it to locate needed libraries.

Signed-off-by: Daniel Mikusa <dan@mikusa.com>